requirements_chatgpt.txt
absl-py==2.2.2
Adafruit-Blinka==8.56.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-motor==3.4.15
adafruit-circuitpython-pca9685==3.4.17
adafruit-circuitpython-register==1.10.2
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-servokit==1.3.19
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
binho-host-adapter==0.1.6
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
Cython==0.29.37
decorator==5.2.1
distro==1.9.0
evdev==1.9.1
filelock==3.18.0
flatbuffers==20181003210633
frozenlist==1.5.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.32.0
google-cloud-texttospeech==2.26.0
google-generativeai==0.8.4
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
lazy_loader==0.4
librosa==0.11.0
lightning-utilities==0.14.3
llvmlite==0.44.0
Markdown==3.8
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
networkx==3.4.2
numba==0.61.2
numpy==2.2.4
onnxruntime==1.21.1
openai==1.75.0
packaging==24.2
piper-phonemize==1.1.0
platformdirs==4.3.7
playsound==1.3.0
pooch==1.8.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pvporcupine==3.0.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyDeprecate==0.3.2
pyftdi==0.56.0
pyparsing==3.2.3
pyserial==3.5
python-dotenv==1.1.0
pytorch-lightning==1.7.7
pyttsx3==2.98
pyusb==1.3.1
PyYAML==6.0.2
requests==2.32.3
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
SpeechRecognition==3.14.2
sympy==1.13.1
sysv-ipc==1.1.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.6.0
torchmetrics==1.7.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.19.0
